Rename random to rng
