+++ src/lib.rs
@@ -36,6 +36,7 @@ pub fn init(boot_info: &'static BootInfo) {
     sys::mem::init(boot_info);
     sys::acpi::init(); // Require MEM
     sys::cpu::init();
+    sys::rng::init();
     sys::pci::init(); // Require MEM
     sys::net::init(); // Require PCI
     sys::ata::init();
