+++ src/sys/mod.rs
@@ -54,7 +54,7 @@ pub mod net;
 pub mod pci;
 pub mod pic;
 pub mod process;
-pub mod random;
+pub mod rng;
 pub mod serial;
 pub mod syscall;
 pub mod time;
